requirements-old.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
confusable-homoglyphs==3.0.0
Django==2.0.4
django-dynamic-raw-id==2.4
django-extensions==2.0.7
django-jinja==2.4.1
django-modelcluster==4.1
django-pagedown==1.0.4
-e git+https://github.com/ubernostrum/django-registration.git@f957d2824cf77820287da2993fca44d6445c592d#egg=django_registration
django-reversion==2.0.13
django-taggit==0.22.2
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.0.0
drf-ujson==1.2
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
Jinja2==2.10
markdown2==2.3.5
MarkupSafe==1.0
Pillow==5.1.0
psycopg2-binary==2.7.4
pysolr==3.7.0
pytz==2018.4
requests==2.18.4
serpy==0.3.1
six==1.11.0
ujson==1.35
Unidecode==0.4.21
urllib3==1.22
wagtail==2.0.1
webencodings==0.5.1
Werkzeug==0.14.1
Willow==1.1
